\name{Internal functions}
\alias{calc2}
\alias{calc3}
\alias{concatenate}
\alias{dbb1}
\alias{dbb2}
\alias{dbb3}
\alias{dbb4}
\alias{degencal2}
\alias{degencal3}
\alias{degencal4}
\alias{exportglobal}
\alias{glob2d}
\alias{glob3d}
\alias{impr2d}
\alias{impr3d}
\alias{OnOk}
\alias{oppdf}
\alias{optxt}
\alias{pcrcal2}
\alias{pcrcal3}
\alias{pcrcal4}
\alias{security}
\title{Internal functions for the Oligo Specificity System package}
\description{Internal functions for the Oligo Specificity System package}
\details{
These are not to be called by the user
}

\author{Rory Michelland \email{rory.michelland@gmail.com}\cr
Laurent Cauquil \email{laurent.cauquil@toulouse.inra.fr}}
\seealso{\code{\link{OSS}}}