\docType{package}
\name{Oligo Specificity System}
\alias{OSS}
\alias{Oss}
\alias{oss}
\alias{OligoSpecificitySystem}
\alias{oligospecificitysystem}
\alias{Oligospecificitysystem}
\title{Oligo Specificity System}
\description{A Tcl/Tk GUI to use functions of the \code{OSS} package.}
\author{Rory Michelland \email{rory.michelland@gmail.com}\cr
Laurent Cauquil \email{laurent.cauquil@toulouse.inra.fr}}

\examples{
\dontrun{
## One of these lines allows the Oligo Specificity System program to start
OSS()
Oss()
oss()
OligoSpecificitySystem()
oligospecificitysystem()
Oligospecificitysystem()
}
}
